socketbuffer.cxx: Add cast to avoid size_t -> unsigned int conversion warning.

+++ log4cplus/src/socketbuffer.cxx
@@ -307,7 +307,7 @@ log4cplus::helpers::SocketBuffer::appendString(const tstring& str)
         return;
     }
 
-    appendInt(strlen);
+    appendInt(static_cast<unsigned>(strlen));
     std::memcpy(&buffer[pos], str.data(), strlen);
     pos += strlen;
     size = pos;
